+++ packages/components-angular/src/app/pages/toast.component.ts
@@ -1,0 +1,46 @@
+/* Auto Generated File */
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { componentsReady, ToastManager } from '@porsche-design-system/components-angular';
+
+@Component({
+  selector: 'page-toast',
+  template: `
+    <div class="visualize-grid">
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+      <span></span>
+    </div>
+
+    <div class="playground light" title="should render toast" style="height: 300px">
+      <p-toast></p-toast>
+    </div>
+  `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class ToastComponent implements OnInit {
+  public allReady: boolean = false;
+
+  constructor(private cdr: ChangeDetectorRef) {}
+
+  ngOnInit() {
+    componentsReady().then(() => {
+      this.allReady = true;
+      this.cdr.markForCheck();
+    });
+  }
+}