+++ packages/components-js/tests/vrt/puppeteer/specs/content-wrapper.vrt.ts
@@ -1,5 +1,0 @@
-import { furtherExtendedViewports, getVisualRegressionTester, vrtTest } from '@porsche-design-system/shared/testing';
-
-it.each(furtherExtendedViewports)('should have no visual regression for viewport %s', async (viewport) => {
-  expect(await vrtTest(getVisualRegressionTester(viewport), 'content-wrapper', '/#content-wrapper')).toBeFalsy();
-});