+++ packages/components/src/components/accordion/accordion-styles.ts
@@ -8,6 +8,8 @@ import {
   addImportantToEachRule,
   hostHiddenStyles,
   hoverMediaQuery,
+  prefersColorSchemeDarkMediaQuery,
+  colorSchemeStyles,
 } from '../../styles';
 import {
   fontWeightSemiBold,
@@ -28,6 +30,12 @@ export const getComponentCss = (
   theme: Theme
 ): string => {
   const { primaryColor, hoverColor, focusColor, contrastLowColor } = getThemedColors(theme);
+  const {
+    primaryColor: primaryColorDark,
+    hoverColor: hoverColorDark,
+    focusColor: focusColorDark,
+    contrastLowColor: contrastLowColorDark,
+  } = getThemedColors('dark');
 
   return getCss({
     '@global': {
@@ -35,7 +43,11 @@ export const getComponentCss = (
         display: 'block',
         ...(!compact && {
           borderBottom: `1px solid ${contrastLowColor}`,
+          ...prefersColorSchemeDarkMediaQuery(theme, {
+            borderColor: contrastLowColorDark,
+          }),
         }),
+        ...colorSchemeStyles,
         ...hostHiddenStyles,
       }),
       button: {
@@ -51,6 +63,9 @@ export const getComponentCss = (
         cursor: 'pointer',
         textAlign: 'left',
         color: primaryColor,
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          color: primaryColorDark,
+        }),
         ...textSmallStyle,
         fontWeight: fontWeightSemiBold,
         ...buildResponsiveStyles(size, (s: AccordionSize) => ({
@@ -83,11 +98,17 @@ export const getComponentCss = (
             },
             '&:hover::before': {
               background: hoverColor,
+              ...prefersColorSchemeDarkMediaQuery(theme, {
+                background: hoverColorDark,
+              }),
             },
           })
         ),
         '&:focus::before': {
           border: `${borderWidthBase} solid ${focusColor}`,
+          ...prefersColorSchemeDarkMediaQuery(theme, {
+            borderColor: focusColorDark,
+          }),
         },
         '&:not(:focus-visible)::before': {
           border: 0,
@@ -113,6 +134,9 @@ export const getComponentCss = (
     },
     collapsible: {
       color: primaryColor, // enables color inheritance for slotted content
+      ...prefersColorSchemeDarkMediaQuery(theme, {
+        color: primaryColorDark,
+      }),
       display: 'grid',
       ...(open
         ? {