+++ packages/components/src/components/fieldset/fieldset-styles.ts
@@ -2,7 +2,13 @@ import type { Theme } from '../../types';
 import type { FieldsetLabelSize } from './fieldset-utils';
 import type { FormState } from '../../utils/form/form-state';
 import { getCss, mergeDeep } from '../../utils';
-import { addImportantToEachRule, getThemedColors, hostHiddenStyles } from '../../styles';
+import {
+  addImportantToEachRule,
+  colorSchemeStyles,
+  getThemedColors,
+  hostHiddenStyles,
+  prefersColorSchemeDarkMediaQuery,
+} from '../../styles';
 import { headingSmallStyle, headingMediumStyle, spacingStaticMedium } from '@porsche-design-system/utilities-v2';
 import { getFunctionalComponentRequiredStyles } from '../common/required/required-styles';
 import { getFunctionalComponentStateMessageStyles } from '../common/state-message/state-message-styles';
@@ -17,6 +23,7 @@ export const getComponentCss = (
     '@global': {
       ':host': addImportantToEachRule({
         display: 'block',
+        ...colorSchemeStyles,
         ...hostHiddenStyles,
       }),
       fieldset: {
@@ -29,6 +36,9 @@ export const getComponentCss = (
           margin: `0 0 ${spacingStaticMedium}`,
           padding: 0,
           color: getThemedColors(theme).primaryColor,
+          ...prefersColorSchemeDarkMediaQuery(theme, {
+            color: getThemedColors('dark').primaryColor,
+          }),
           ...(labelSize === 'small' ? headingSmallStyle : headingMediumStyle),
         },
       }),