+++ packages/components/src/components/inline-notification/inline-notification-styles.ts
@@ -2,7 +2,13 @@ import type { Theme } from '../../types';
 import type { InlineNotificationState } from './inline-notification-utils';
 import { getMediaQueryMax, headingSmallStyle, textSmallStyle } from '@porsche-design-system/utilities-v2';
 import { getCss } from '../../utils';
-import { addImportantToEachRule, getThemedColors, hostHiddenStyles } from '../../styles';
+import {
+  addImportantToEachRule,
+  colorSchemeStyles,
+  getThemedColors,
+  hostHiddenStyles,
+  prefersColorSchemeDarkMediaQuery,
+} from '../../styles';
 import {
   getNotificationContentJssStyle,
   getNotificationIconJssStyle,
@@ -21,6 +27,7 @@ export const getComponentCss = (
     '@global': {
       ':host': addImportantToEachRule({
         ...getNotificationRootJssStyle(state, hasAction, hasClose, theme),
+        ...colorSchemeStyles,
         ...hostHiddenStyles,
       }),
       h5: headingSmallStyle,
@@ -28,6 +35,9 @@ export const getComponentCss = (
       'h5,p': {
         margin: 0,
         color: getThemedColors(theme).primaryColor,
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          color: getThemedColors('dark').primaryColor,
+        }),
       },
     },
     icon: getNotificationIconJssStyle(),