+++ packages/components/src/components/stepper-horizontal/stepper-horizontal/stepper-horizontal-styles.ts
@@ -2,14 +2,17 @@ import { buildResponsiveStyles, getCss } from '../../../utils';
 import type { StepperHorizontalSize } from './stepper-horizontal-utils';
 import type { BreakpointCustomizable } from '../../../types';
 import { fontSizeText, textSmallStyle } from '@porsche-design-system/utilities-v2';
-import { addImportantToEachRule, hostHiddenStyles } from '../../../styles';
+import { addImportantToEachRule, colorSchemeStyles, hostHiddenStyles } from '../../../styles';
 
 export const getComponentCss = (size: BreakpointCustomizable<StepperHorizontalSize>): string => {
   return getCss({
     '@global': {
       ':host': {
         display: 'block',
-        ...addImportantToEachRule(hostHiddenStyles),
+        ...addImportantToEachRule({
+          ...colorSchemeStyles,
+          ...hostHiddenStyles,
+        }),
       },
     },
     scroller: {