+++ packages/components/src/styles/index.ts
@@ -2,5 +2,7 @@
 export * from './colors';
 export * from './common-styles';
 export * from './host-hidden-styles';
+export * from './color-scheme-styles';
 export * from './hover-media-query';
+export * from './prefers-color-scheme-dark-media-query';
 export * from './schemed-high-contrast-media-query';