+++ packages/components/src/styles/tile/tile-base-styles.ts
@@ -2,7 +2,14 @@ import type { Styles } from 'jss';
 import type { BreakpointCustomizable } from '../../types';
 import type { TileAspectRatio } from '../../utils';
 import { buildResponsiveStyles } from '../../utils';
-import { addImportantToEachRule, getInsetJssStyle, getTransition, hostHiddenStyles, hoverMediaQuery } from '../';
+import {
+  addImportantToEachRule,
+  colorSchemeStyles,
+  getInsetJssStyle,
+  getTransition,
+  hostHiddenStyles,
+  hoverMediaQuery,
+} from '../';
 import { borderRadiusLarge, spacingStaticMedium } from '@porsche-design-system/utilities-v2';
 import { getThemedTypographyColor } from '../text-icon-styles';
 
@@ -23,7 +30,10 @@ export const getTileBaseStyles = (
       ':host': {
         display: 'block',
         hyphens: 'auto',
-        ...addImportantToEachRule(hostHiddenStyles),
+        ...addImportantToEachRule({
+          ...colorSchemeStyles,
+          ...hostHiddenStyles,
+        }),
       },
       ...addImportantToEachRule({
         '::slotted': {