+++ packages/components/src/utils/index.ts
@@ -18,7 +18,6 @@ export * from './is-click-outside';
 export * from './is-ssr-hydration';
 export * from './jss';
 export { parseJSONAttribute } from './json';
-export * from './link-button/getLinkButtonThemeForIcon';
 export * from './link-button-tile/link-button-tile-utils';
 export * from './tile/tile-utils';
 export * from './log';